requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
confusable-homoglyphs==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-cors-headers==3.1.1
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
entrypoints==0.3
flake8==3.7.8
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==6.2.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.6